GCloud API wrapper
This PR adds the wrapper for Google Cloud Storage SDK as well
as a simple fake implementation.

Unfortunately, there is no Storage mock API that supports RPC,
hence we'll need a fake for other tests.

There is a good chance that the fake will change as we use it
but i still wanted to merge one with this CL to begin with