ApkStore
This PR adds a new class that is responsible for pushing APKs to GCS.

It is fairly primitive such that it will upload the apk to GCS if it doese not exists.
We use a composite key from apk contents + name to compute a unique key.

I've not tested what happens if GCP uploads race so there is a possibility that
if two runners run at the same time w/ the same APK things might go wrong.
We'll deal with it if that happens. As long as GCP does not consider an Object
exists until it is fully uploaded, it should not be a problem.