Add TestMatrixStore
This is a wrapper around the firebase APIs to track test matrices.
For each test run, we create a unique key based on the hashes of the
apks + test environment.
Using this key, we associate the test run with a TestMatrixId such that
if same parameters are passed again, we can return the existing TestMatrix
instead of a new one.