Add support for running integration tests for libraries
FTL always requires a test apk, hence we cannot run instrumentation
tests for libraries.

This CL adds an android module which gets compiled into resources to
be used as a placeholder apk when we don't have one.

Unfortunately, FTL also uses that APKs package to decide on the History
name. To avoid having all library tests under 1 name, I've implemented
integration with the tools reseult API to create a history. Luckily, I
over engineered when we needed the FTL API so i was able to re-use that to
generate models for the ToolsResult API. For now, we use the test apk name
as the history name which we may want to change later. This also means we'll
duplicate all histories (they use package name). There is no API to delete them
but I'm hoping that aafter 90 days, they'll disappear.