Use test package name for history id on firebase
Previously, we were using test apk names for history name in
FTL. That name can change (based on androidX) and it might also
be too long (100 chars). Instead, this CL replaces it with the
package name of the test apk (using an FTL service). This does
mess up the FTL UI for 3 months but afterwards, should be much
easier to read and much more stable.

Also updated kotlin to 1.5.30