Pick device version from supported versions
This PR is a quick fix to pick a proper device/version pair.

Previously, we picked default device and default version but it does not
guarantee that the default device supports the default version.

This PR changes that logic to pick the max version from the default
device.
This is a band-aid, we should rewrite this section to pick a device
based on the targetSdk of the APK.

Test: FirebaseTestLabControllerTest