Expand CI action to receive all necessary arguments as parameters (#37)
* Expand CLI ci-action (#36)

* Expand CLI ci-action

This CL expands the CI action to receive more parameters and hardcode
less values to be more usable from outside.

Added log-file(optional) to let caller specify a log file location outside the
output directory.

Added device-specs to be able to pass list of device-api configurations
as an argument.

Added artifact-name-regex (optional) that can be used to filter artifacts
by name. The list of artifacts from a github build might be many and downloading
all of them to look for APKs is silly. This parameter can be optimized to
reduce the # of artifacts that are searched.

Added gcp-bucket-name to specify the gcp butket name as an argument, instead
of the previously hardcoded value.

Added gcp-bucket-path to specify the path in the bucket, instead of the
previously hardcoded value.

* cleanup

* make device list optional

* ktformat

* ignore empty artifact filter regex

* remove unused variable