Add parameter to ignore fully skipped tests
We have some test artifacts that have no non-skipped tests. This CL adds a new
action parameter that allows ignoring them.

When all test cases in an APK are skipped, FTL marks its outcome as FAILED instead
of SKIPPED. In this CL, we fix it back to SKIPPED if requested.

Test: TestRunnerTest